error → debug in log messages before raising exceptions
The messages are useful because they indicate the request URL,
which is not part of the original exception, but we do not want
to log an error message if the exception is eventually handled,
e.g. by a middleware or errback, and if not, the exception itself
will trigger an error message.